+++ package.json
@@ -16,8 +16,11 @@
     "history": "^5.0.0",
     "jquery": "^3.5.1",
     "react": "^17.0.1",
+    "react-awesome-lightbox": "^1.7.3",
+    "react-bnb-gallery": "^1.4.4",
     "react-bootstrap": "^1.4.0",
     "react-dom": "^17.0.1",
+    "react-image-gallery": "^1.0.8",
     "react-router-dom": "^5.2.0",
     "react-scripts": "^4.0.1",
     "react-toastify": "^6.1.0",
